+++ actionpack/lib/action_dispatch/testing/integration.rb
@@ -80,59 +80,6 @@ module ActionDispatch
         get(response.location)
         status
       end
-
-      # Performs a request using the specified method, following any subsequent
-      # redirect. Note that the redirects are followed until the response is
-      # not a redirect--this means you may run into an infinite loop if your
-      # redirect loops back to itself.
-      #
-      # Example:
-      #
-      #   request_via_redirect :post, '/welcome',
-      #     params: { ref_id: 14 },
-      #     headers: { "X-Test-Header" => "testvalue" }
-      def request_via_redirect(http_method, path, *args)
-        ActiveSupport::Deprecation.warn("`request_via_redirect` is deprecated and will be removed in Rails 5.1. Please use `follow_redirect!` manually after the request call for the same behavior.")
-        process_with_kwargs(http_method, path, *args)
-
-        follow_redirect! while redirect?
-        status
-      end
-
-      # Performs a GET request, following any subsequent redirect.
-      # See +request_via_redirect+ for more information.
-      def get_via_redirect(path, *args)
-        ActiveSupport::Deprecation.warn("`get_via_redirect` is deprecated and will be removed in Rails 5.1. Please use `follow_redirect!` manually after the request call for the same behavior.")
-        request_via_redirect(:get, path, *args)
-      end
-
-      # Performs a POST request, following any subsequent redirect.
-      # See +request_via_redirect+ for more information.
-      def post_via_redirect(path, *args)
-        ActiveSupport::Deprecation.warn("`post_via_redirect` is deprecated and will be removed in Rails 5.1. Please use `follow_redirect!` manually after the request call for the same behavior.")
-        request_via_redirect(:post, path, *args)
-      end
-
-      # Performs a PATCH request, following any subsequent redirect.
-      # See +request_via_redirect+ for more information.
-      def patch_via_redirect(path, *args)
-        ActiveSupport::Deprecation.warn("`patch_via_redirect` is deprecated and will be removed in Rails 5.1. Please use `follow_redirect!` manually after the request call for the same behavior.")
-        request_via_redirect(:patch, path, *args)
-      end
-
-      # Performs a PUT request, following any subsequent redirect.
-      # See +request_via_redirect+ for more information.
-      def put_via_redirect(path, *args)
-        ActiveSupport::Deprecation.warn("`put_via_redirect` is deprecated and will be removed in Rails 5.1. Please use `follow_redirect!` manually after the request call for the same behavior.")
-        request_via_redirect(:put, path, *args)
-      end
-
-      # Performs a DELETE request, following any subsequent redirect.
-      # See +request_via_redirect+ for more information.
-      def delete_via_redirect(path, *args)
-        ActiveSupport::Deprecation.warn("`delete_via_redirect` is deprecated and will be removed in Rails 5.1. Please use `follow_redirect!` manually after the request call for the same behavior.")
-        request_via_redirect(:delete, path, *args)
-      end
     end
 
     # An instance of this class represents a set of requests and responses